requirements.txt
whoosh==2.7.4
numpy==1.23.5
PyYAML==6.0.1
retry==0.9.2
nltk==3.8.1
ftfy==6.1.1
beautifulsoup4==4.12.3
tqdm==4.66.4
transformers==4.41.2
torch==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
scipy==1.10.1
together==0.2.11
